+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.5.1 - 2024-0x-xx]
+
+### Added
+
+- Test deploy button in Admin settings for each Daemon configuration.
+
 ## [2.5.0 - 2024-04-23]
 
 ### Added